update the copyright
add a lot of comments at the beginning to explain things.  these
were sorely lacking.

fix the sigaction stuff to not output warnings.

add a macro to define the maximum run period.  possibly someone might
like a different run period, like days or something, if not storing
the readings to a file and willing to use a lot of memory.